refactor into faculty loads service
